+++ cardano-tracer/src/Cardano/Tracer/Configuration.hs
@@ -77,7 +77,8 @@ instance FromJSON RotationParams where
     rpFrequencySecs <- o .: "rpFrequencySecs"
     rpLogLimitBytes <- o .: "rpLogLimitBytes"
     rpMaxAgeMinutes <- o .: "rpMaxAgeMinutes"
-                   <|> o .: "rpMaxAgeHours" <&> (* 60)
+                   <|> (o .: "rpMaxAgeHours" <&> (* 60))
+                   <|> pure (24 * 60)
     rpKeepFilesNum  <- o .: "rpKeepFilesNum"
     pure RotationParams{..}