+++ cardano-tracer/src/Cardano/Tracer/Utils.hs
@@ -39,7 +39,7 @@ module Cardano.Tracer.Utils
   , sequenceConcurrently_
   ) where
 
-import           Cardano.Node.Startup (NodeInfo (..))
+import           Cardano.Logging.Types.NodeInfo (NodeInfo(..))
 import           Cardano.Tracer.Configuration
 import           Cardano.Tracer.Environment
 import           Cardano.Tracer.Handlers.Utils