+++ trace-dispatcher/src/Cardano/Logging/Types.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
@@ -105,7 +106,7 @@ instance Monad m => Monoid (Trace m a) where
 data Namespace a = Namespace {
     nsPrefix :: [Text]
   , nsInner  :: [Text]}
-  deriving Eq
+  deriving stock Eq
 
 instance Show (Namespace a) where
   show (Namespace [] []) = "emptyNS"
@@ -210,7 +211,7 @@ data Metric
   -- the metric will be represented as "prometheus_metric{key1=\"value1\",key2=\"value2\"} 1"
 
     | PrometheusM Text [(Text, Text)]
-  deriving (Show, Eq)
+  deriving stock (Eq, Show)
 
 
 getMetricName :: Metric -> Text
@@ -229,10 +230,8 @@ emptyObject = HM.empty
 -- important to provide a complete list, as the prototypes are used as well for configuration.
 -- If you don't want to add an item for documentation enter an empty text.
 newtype Documented a = Documented {undoc :: [DocMsg a]}
-  deriving Show
-
-instance Semigroup (Documented a) where
-  (<>) (Documented l) (Documented r) = Documented (l ++ r)
+  deriving stock Show
+  deriving newtype Semigroup
 
 -- | Document a message by giving a prototype, its most special name in the namespace
 -- and a comment in markdown format
@@ -253,7 +252,10 @@ data LoggingContext = LoggingContext {
   , lcPrivacy   :: Maybe Privacy
   , lcDetails   :: Maybe DetailLevel
   }
-  deriving Show
+  deriving stock 
+    (Show, Generic)
+  deriving anyclass
+    Serialise
 
 emptyLoggingContext :: LoggingContext
 emptyLoggingContext = LoggingContext [] [] Nothing Nothing Nothing
@@ -264,17 +266,18 @@ data DetailLevel =
     | DNormal
     | DDetailed
     | DMaximum
-  deriving (Show, Eq, Ord, Bounded, Enum, Generic, Serialise)
+  deriving stock (Eq, Ord, Show, Enum, Bounded, Generic)
+  deriving anyclass (Serialise, AE.FromJSON)
 
 instance AE.ToJSON DetailLevel where
     toEncoding = AE.genericToEncoding AE.defaultOptions
-instance AE.FromJSON DetailLevel
 
 -- | Privacy of a message. Default is Public
 data Privacy =
       Confidential              -- ^ confidential information - handle with care
     | Public                    -- ^ can be public.
-      deriving (Show, Eq, Ord, Bounded, Enum, Generic, Serialise)
+  deriving stock (Eq, Ord, Show, Enum, Bounded, Generic)
+  deriving anyclass Serialise
 
 -- | Severity of a message
 data SeverityS
@@ -286,13 +289,14 @@ data SeverityS
     | Critical                -- ^ Severe situations
     | Alert                   -- ^ Take immediate action
     | Emergency               -- ^ System is unusable
-  deriving (Show, Eq, Ord, Bounded, Enum, Read, AE.ToJSON, AE.FromJSON, Generic, Serialise)
+  deriving stock (Eq, Ord, Show, Read, Enum, Bounded, Generic)
+  deriving anyclass (AE.ToJSON, AE.FromJSON, Serialise)
 
 -- | Severity for a filter
 -- Nothing means don't show anything (Silence)
 -- Nothing level means show messages with severity >= level
 newtype SeverityF = SeverityF (Maybe SeverityS)
-  deriving (Eq)
+  deriving stock Eq
 
 instance Enum SeverityF where
   toEnum 8 = SeverityF Nothing
@@ -351,7 +355,7 @@ data FormattedMessage =
     | FormattedMachine Text
     | FormattedMetrics [Metric]
     | FormattedForwarder TraceObject
-  deriving (Eq, Show)
+  deriving stock (Eq, Show)
 
 
 data PreFormatted = PreFormatted {
@@ -373,7 +377,11 @@ data TraceObject = TraceObject {
   , toTimestamp :: !UTCTime
   , toHostname  :: !Text
   , toThreadId  :: !Text
-} deriving (Eq, Show)
+} deriving stock 
+    (Eq, Show, Generic)
+  -- ^ Instances for 'TraceObject' to forward it using 'trace-forward' library.
+  deriving anyclass
+    (Serialise, ShowProxy)
 
 -- |
 data BackendConfig =
@@ -382,7 +390,7 @@ data BackendConfig =
   | EKGBackend
   | DatapointBackend
   | PrometheusSimple Bool (Maybe HostName) PortNumber   -- boolean: drop suffixes like "_int" in exposition; default: False
-  deriving (Eq, Ord, Show, Generic)
+  deriving stock (Eq, Ord, Show, Generic)
 
 instance AE.ToJSON BackendConfig where
   toJSON Forwarder  = AE.String "Forwarder"
@@ -428,7 +436,7 @@ data FormatLogging =
     HumanFormatColoured
   | HumanFormatUncoloured
   | MachineFormat
-  deriving (Eq, Ord, Show)
+  deriving stock (Eq, Ord, Show)
 
 -- Configuration options for individual namespace elements
 data ConfigOption =
@@ -442,11 +450,11 @@ data ConfigOption =
   -- | Construct a limiter with limiting to the Double,
   -- which represents frequency in number of messages per second
   | ConfLimiter {maxFrequency :: Double}
-  deriving (Eq, Ord, Show, Generic)
+  deriving stock (Eq, Ord, Show, Generic)
 
 newtype ForwarderAddr
   = LocalSocket FilePath
-  deriving (Eq, Ord, Show)
+  deriving stock (Eq, Ord, Show)
 
 instance AE.FromJSON ForwarderAddr where
   parseJSON = AE.withObject "ForwarderAddr" $ \o -> LocalSocket <$> o AE..: "filePath"
@@ -458,14 +466,15 @@ data ForwarderMode =
     -- | Forwarder works as a server: it accepts network connection from
     -- 'cardano-tracer' and/or another Haskell acceptor application.
   | Responder
-  deriving (Eq, Ord, Show, Generic)
+  deriving stock (Eq, Ord, Show, Generic)
 
 data Verbosity =
     -- | Maximum verbosity for all tracers in the forwarding protocols.
     Maximum
     -- | Minimum verbosity, the forwarding will work as silently as possible.
   | Minimum
-  deriving (Eq, Ord, Show, Generic, AE.ToJSON)
+  deriving stock (Eq, Ord, Show, Generic)
+  deriving anyclass AE.ToJSON
 
 instance AE.FromJSON Verbosity where
   parseJSON (AE.String "Maximum") = pure Maximum
@@ -478,7 +487,8 @@ data TraceOptionForwarder = TraceOptionForwarder {
   , tofDisconnQueueSize    :: Word
   , tofVerbosity           :: Verbosity
   , tofMaxReconnectDelay   :: Word
-} deriving (Eq, Generic, Ord, Show, AE.ToJSON)
+} deriving stock (Eq, Ord, Show, Generic)
+  deriving anyclass AE.ToJSON
 
 -- A word regarding queue sizes:
 -- In case of a missing forwarding service consumer, traces messages will be
@@ -533,8 +543,7 @@ data TraceConfig = TraceConfig {
     -- | Optional resource trace frequency in milliseconds.
   , tcResourceFrequency :: Maybe Int
 }
-  deriving (Eq, Ord, Show)
-
+  deriving stock (Eq, Ord, Show)
 
 emptyTraceConfig :: TraceConfig
 emptyTraceConfig = TraceConfig {
@@ -572,7 +581,7 @@ data LogDoc = LogDoc {
   , ldFiltered        :: ![SeverityF]
   , ldLimiter         :: ![(Text, Double)]
   , ldSilent          :: Bool
-} deriving(Eq, Show)
+} deriving stock (Eq, Show)
 
 emptyLogDoc :: Text -> [(Text, Text)] -> LogDoc
 emptyLogDoc d m = LogDoc d (Map.fromList m) [] Nothing Nothing Nothing [] [] [] [] False
@@ -587,14 +596,3 @@ instance LogFormatting b => LogFormatting (Folding a b) where
   forMachine v (Folding b) =  forMachine v b
   forHuman (Folding b)     =  forHuman b
   asMetrics (Folding b)    =  asMetrics b
-
----------------------------------------------------------------------------
--- Instances for 'TraceObject' to forward it using 'trace-forward' library.
-
-deriving instance Generic LoggingContext
-deriving instance Generic TraceObject
-
-instance Serialise LoggingContext
-instance Serialise TraceObject
-
-instance ShowProxy TraceObject