+++ cardano-node/ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Next version
 
+- Removed `cardano-node' as a dependency from `cardano-tracer'. This necessitated moving `NodeInfo` from
+  `cardano-tracer:Cardano.Node.Startup`to `trace-dispatcher:Cardano.Logging.Types.NodeInfo`, and `NodePeers` from
+  `cardano-node:Cardano.Node.Tracing.Peers` to `trace-dispatcher:Cardano.Logging.Types.NodePeers`.
+
 - Add a new configuration field for fork-policy.
 
 - Optionally support lightweight checkpointing.